Move LocalSwapId to the comit crate
In preparation for moving the Comit network behaviour over to the
comit crate; move the `LocalSwapId` type to the comit crate.
